+++ R/lasclip.r
@@ -29,38 +29,56 @@
 #' Clip LiDAR points
 #'
 #' Clip LiDAR points within a given geometry from a point cloud (\code{LAS} object) or a catalog
-#' (\code{LAScatalog} object)
+#' (\code{LAScatalog} object). With a \code{LAS} object, the user first reads and loads a point-cloud
+#' in memory and then can clip it to get a subset  within a region of interest (ROI). With a \code{LAScatalog}
+#' object, the user can extracts any arbitrary ROI for a set of \code{las/laz} file loading only the
+#' points of interest. This is faster, easier and much more memory-efficient for extracting ROIs.
 #'
-#' \code{lasclip} functions work both on \code{LAS} and \code{LAScatalog} objects. With a \code{LAS}
-#' object, the user first reads and loads a point-cloud and then clip it to get a subset within a region
-#' of interest (ROI). With a \code{LAScatalog} object, the user extracts the ROI without
-#' loading the whole point-cloud. This is faster and much more memory-efficient for extracting ROIs.
-#' \cr\cr
-#' Minor differences exist between the \code{LAS} and \code{LAScatalog} version of \code{lasclip}.
-#' For example the user can clip a \code{SpatialPolygonsDataFrame} from a \code{LAS} object but not
-#' from a \code{LAScatalog}. Also the option \code{inside = FALSE} is disabled for \code{LAScatalog}
-#' objects. These differences are generally justified by memory safety concerns.
+#' @section Supported geometries:
+#' \itemize{
+#'  \item \href{https://en.wikipedia.org/wiki/Well-known_text}{WKT string}: describing a POLYGON or
+#'  a MULTIPOLYGON. A single \code{LAS} object is extracted.
+#'  \item \link[sp:Polygon-class]{Polygon}
+#'  \item \link[sp:Polygons-class]{Polygons}
+#'  \item \link[sp:SpatialPolygons-class]{SpatialPolygons}
+#'  \item \link[sp:SpatialPolygonsDataFrame-class]{SpatialPolygonsDataFrame}
+#'  \item \link[raster:Extent-class]{Extent}: represent a bouding box. A rectangle is extracted.
+#'  \item \link[base:matrix]{matrix} 2 x 2 describing a bounding box following this order:
+#'  \preformatted{
+#'   min     max
+#' x 684816  684943
+#' y 5017823 5017957}
+#'  \item Any other object that have a bouding box accessible via \code{raster::extent} such as a
+#'  \link[raster:RasterLayer-class]{RasterLayer} are supported. A rectangle is extracted.
+#' }
 #'
-#' @param x An object of class \code{LAS} or \code{LAScatalog}.
-#' @param geometry a geometric object. Currently \code{Polygon} and \code{SpatialPolygonsDataFrame}
-#' from \code{sp} are supported.
-#' @param xleft scalar of left x position of rectangle.
-#' @param ybottom	scalar of bottom y position of rectangle.
-#' @param xright scalar of right x position of rectangle.
-#' @param ytop scalar of top y position of rectangle.
-#' @param xpoly numerical array. x-coordinates of polygon.
-#' @param ypoly numerical array. y-coordinates of polygon.
-#' @param xcenter scalar of x disc center.
-#' @param ycenter scalar of y disc center.
-#' @param radius scalar of disc radius.
-#' @param ofile character. Path to an output file (only with a \code{LAScatalog} object).
-#' If \code{ofile = ""} the result is loaded into R, otherwise the result is written to a
-#' file while reading. This is much more memory-efficient than loading into R first, then writing.
-#' @param inside logical. Inverts the selection (only with a \code{LAS} object). Select inside or outside
-#' the shape.
-#' @param ... Additional argument for readLAS to reduce the amount of data loaded (only with a
+#' @template LAScatalog
+#'
+#' @section Supported processing options for a LAScatalog:
+#' \itemize{
+#' \item \code{cores}: If several ROIs are requested, they can be extracted in parallel.
+#' \item \code{progress}: See \link{LAScatalog-class}.
+#' \item \code{stop_early}: See \link{LAScatalog-class}.
+#' \item \code{save}: If save is set in the catalog, the ROIs will no be returns in R. They will be written
+#' in on or several files. See \link{LAScatalog-class} and example.
+#' }
+#'
+#' @template param-las
+#' @param geometry a geometric object. Many types are supported, see section 'supported geometries'.
+#' @param xleft numeric. left x coordinates of rectangles.
+#' @param ybottom	numeric. bottom y coordinates of rectangles.
+#' @param xright numeric. right x coordinates of rectangles.
+#' @param ytop numeric. top y coordinates of rectangles.
+#' @param xpoly numeric. x coordinates of a polygon.
+#' @param ypoly numeric. y coordinates of a polygon.
+#' @param xcenter numeric. x coordinates of discs centers.
+#' @param ycenter numeric. y coordinates of discs centers.
+#' @param radius numeric. disc radiuses.
+#' @param ... Additional argument for \link{readLAS} to reduce the amount of data loaded (only with a
 #' \code{LAScatalog} object)
-#' @return An object of class \code{LAS} or NULL if the result is immediately written to a file.
+#' @return An object of class \code{LAS} or a \code{list} of \code{LAS} objects if the query implies to return
+#' several regions of interest or \code{NULL} if the query is outside the dataset or within a region that does
+#' not contains any point.
 #' @examples
 #' LASfile <- system.file("extdata", "Megaplot.laz", package="lidR")
 #'
@@ -80,7 +98,7 @@
 #'
 #' # Extract a polygon, write it in a file, do not load anything in R
 #' file = paste0(tempfile(), ".las")
-#' lasclip(ctg, lake, ofile = file)
+#' lasclip(ctg, lake)
 #'
 #' \dontrun{
 #' plot(subset1)
@@ -89,48 +107,60 @@
 #' }
 #' @name lasclip
 #' @export
-#' @export
-lasclip = function(x, geometry, ofile = "", inside = TRUE, ...)
+lasclip = function(las, geometry, ...)
 {
-  UseMethod("lasclip", x)
-}
-
-#' @export
-lasclip.LAS = function(x, geometry, ofile = "", inside = TRUE, ...)
-{
-  assertive::assert_is_a_string(ofile)
-  assertive::assert_is_a_bool(inside)
+  if (is.character(geometry))
+    return(lasclipWKT(las, geometry, ...))
 
   if (is(geometry, "Polygon"))
+    geometry <- sp::Polygons(list(geometry), ID = 1)
+
+  if (is(geometry, "Polygons"))
+    geometry <- sp::SpatialPolygons(list(geometry))
+
+  if (is(geometry, "SpatialPolygons") | is(geometry, "SpatialPolygonsDataFrame"))
+    geometry <- sf::st_as_sf(geometry)
+
+  if (is(geometry, "sf"))
   {
-     las = lasclipPolygon(x, geometry@coords[,1], geometry@coords[,2], inside = inside)
-     return(las)
+    if (!all(sf::st_is(geometry, "POLYGON") |sf::st_is(geometry, "MULTIPOLYGON")))
+      stop("Incorrect geometry type. POLYGON and MULTIPOLYGON are supported.", call. = FALSE)
+
+    geometry <- sf::st_as_text(geometry$geometry)
+    return(lasclipWKT(las, geometry, ...))
   }
-  else if (is(geometry, "SpatialPolygonsDataFrame"))
+  else if (is(geometry, "Extent"))
   {
-    id = classify_from_shapefile(x, geometry)
-    X = split(x@data, id)
-    X = lapply(X, LAS, header = x@header)
-    return(X)
+    xmin = geometry@xmin
+    xmax = geometry@xmax
+    ymin = geometry@ymin
+    ymax = geometry@ymax
+    return(lasclipRectangle(las, xmin, ymin, xmax, ymax, ...))
+  }
+  else if (is.matrix(geometry))
+  {
+    if (!all(dim(geometry) == 2))
+      stop("Matrix must have a size 2 x 2", call. = FALSE)
+
+    xmin = geometry[1]
+    xmax = geometry[3]
+    ymin = geometry[2]
+    ymax = geometry[4]
+    return(lasclipRectangle(las, xmin, ymin, xmax, ymax, ...))
+  }
+  else if (tryCatch(is(raster::extent(geometry), "Extent"), error = function(e) return(FALSE)))
+  {
+    geometry = raster::extent(geometry)
+    xmin = geometry@xmin
+    xmax = geometry@xmax
+    ymin = geometry@ymin
+    ymax = geometry@ymax
+    return(lasclipRectangle(las, xmin, ymin, xmax, ymax, ...))
   }
   else
   {
-    stop("Geometry not supported", call. = FALSE)
+    stop(paste0("Geometry type ", paste0(class(geometry), collapse = " "), " not supported"), call. = FALSE)
   }
-}
-
-#' @export
-lasclip.LAScatalog = function(x, geometry, ofile = "", inside = TRUE, ...)
-{
-  assertive::assert_is_a_string(ofile)
-  assertive::assert_is_a_bool(inside)
-
-  if (is(geometry, "Polygon"))
-    return(lasclipPolygon(x, geometry@coords[,1], geometry@coords[,2], ofile, inside))
-  else if (is(geometry, "SpatialPolygonsDataFrame"))
-    stop("'SpatialPolygonsDataFrame' is not supported for 'LAScatalog' objects", call. = FALSE)
-  else
-    stop("Geometry not supported", call. = FALSE)
 }
 
 # =========
@@ -139,13 +169,13 @@ lasclip.LAScatalog = function(x, geometry, ofile = "", inside = TRUE, ...)
 
 #' @export
 #' @rdname lasclip
-lasclipRectangle = function(x, xleft, ybottom, xright, ytop, ofile = "", inside = TRUE, ...)
+lasclipRectangle = function(las, xleft, ybottom, xright, ytop, ...)
 {
-  UseMethod("lasclipRectangle", x)
+  UseMethod("lasclipRectangle", las)
 }
 
 #' @export
-lasclipRectangle.LAS = function(x, xleft, ybottom, xright, ytop, ofile = "", inside = TRUE, ...)
+lasclipRectangle.LAS = function(las, xleft, ybottom, xright, ytop, ...)
 {
   assertive::assert_is_numeric(xleft)
   assertive::assert_is_numeric(ybottom)
@@ -154,35 +184,33 @@ lasclipRectangle.LAS = function(x, xleft, ybottom, xright, ytop, ofile = "", ins
   assertive::assert_are_same_length(xleft, ybottom)
   assertive::assert_are_same_length(xleft, xright)
   assertive::assert_are_same_length(xleft, ytop)
-  assertive::assert_is_a_string(ofile)
-  assertive::assert_is_a_bool(inside)
 
   X <- Y <- NULL
 
-  if (length(xleft) == 1)
+  output = vector(mode = "list", length(xleft))
+  for (i in 1:length(xleft))
   {
-    if (inside)
-      return(lasfilter(x, between(X, xleft, xright), between(Y, ybottom, ytop)))
-    else
-      return(lasfilter(x, !(between(X, xleft, xright) & between(Y, ybottom, ytop))))
-  }
-  else
-  {
-    output = vector(mode = "list", length(xleft))
-    for (i in 1:length(xleft))
+    roi =  lasfilter(las, X >= xleft[i] & X < xright[i] & Y >= ybottom[i] & Y < ytop[i])
+    if (is.null(roi))
     {
-      if (inside)
-        output[[i]] = lasfilter(x, between(X, xleft[i], xright[i]), between(Y, ybottom[i], ytop[i]))
-      else
-        output[[i]] = lasfilter(x, !(between(X, xleft[i], xright[i]) & between(Y, ybottom[i], ytop[i])))
+      warning(glue::glue("No point found for within disc ({xleft[i]}, {ybottom[i]}, {xright[i]}, {ytop[i]}). NULL returned."), call. = FALSE)
+      output[i] = list(NULL)
     }
-
-    return(output)
+    else
+      output[[i]] = roi
   }
+
+
+  if(length(output) == 0)
+    return(NULL)
+  else if (length(output) == 1)
+    return(output[[1]])
+  else
+    return(output)
 }
 
 #' @export
-lasclipRectangle.LAScatalog = function(x, xleft, ybottom, xright, ytop, ofile = "", inside = TRUE, ...)
+lasclipRectangle.LAScatalog = function(las, xleft, ybottom, xright, ytop, ...)
 {
   assertive::assert_is_numeric(xleft)
   assertive::assert_is_numeric(ybottom)
@@ -191,23 +219,16 @@ lasclipRectangle.LAScatalog = function(x, xleft, ybottom, xright, ytop, ofile = 
   assertive::assert_are_same_length(xleft, ybottom)
   assertive::assert_are_same_length(xleft, xright)
   assertive::assert_are_same_length(xleft, ytop)
-  assertive::assert_is_a_string(ofile)
-  assertive::assert_is_a_bool(inside)
 
-  if (!inside) stop("'inside = FALSE' is not supported for 'LAScatalog' objects.")
+  bboxes  = mapply(raster::extent, xleft, xright, ybottom, ytop)
+  output  = catalog_extract(las, bboxes, LIDRRECTANGLE, ...)
 
-  if (length(xleft) == 1)
-  {
-    return(catalog_clip_rect(x, xleft, ybottom, xright, ytop, ofile, ...))
-  }
+  if(length(output) == 0)
+    return(NULL)
+  else if (length(output) == 1)
+    return(output[[1]])
   else
-  {
-    xcenter = (xleft + xright)/2
-    ycenter = (ybottom + ytop)/2
-    width   = (xright - xleft)/2
-    height  = (ytop - ybottom)/2
-    return(catalog_queries(x, xcenter, ycenter, width, height, ...))
-  }
+    return(output)
 }
 
 # ========
@@ -216,40 +237,14 @@ lasclipRectangle.LAScatalog = function(x, xleft, ybottom, xright, ytop, ofile = 
 
 #' @export lasclipPolygon
 #' @rdname lasclip
-lasclipPolygon = function(x, xpoly, ypoly, ofile = "", inside = TRUE, ...)
-{
-  UseMethod("lasclipPolygon", x)
-}
-
-#' @export
-lasclipPolygon.LAS = function(x, xpoly, ypoly, ofile = "", inside = TRUE, ...)
+lasclipPolygon = function(las, xpoly, ypoly, ...)
 {
   assertive::assert_is_numeric(xpoly)
   assertive::assert_is_numeric(ypoly)
   assertive::assert_are_same_length(xpoly, ypoly)
-  assertive::assert_is_a_string(ofile)
-  assertive::assert_is_a_bool(inside)
 
-  X <- Y <- NULL
-
-  if( inside)
-    return(lasfilter(x, C_points_in_polygon(xpoly,ypoly, X, Y)))
-  else
-    return(lasfilter(x, !C_points_in_polygon(xpoly,ypoly, X, Y)))
-}
-
-#' @export
-lasclipPolygon.LAScatalog = function(x, xpoly, ypoly, ofile = "", inside = TRUE, ...)
-{
-  assertive::assert_is_numeric(xpoly)
-  assertive::assert_is_numeric(ypoly)
-  assertive::assert_are_same_length(xpoly, ypoly)
-  assertive::assert_is_a_string(ofile)
-  assertive::assert_is_a_bool(inside)
-
-  if (!inside) stop("'inside = FALSE' is not supported for 'LAScatalog' objects.")
-
-  return(catalog_clip_poly(x, xpoly, ypoly, ofile, ...))
+  poly = sp::Polygon(cbind(xpoly, ypoly))
+  return(lasclip(las, poly, ...))
 }
 
 # ========
@@ -258,171 +253,189 @@ lasclipPolygon.LAScatalog = function(x, xpoly, ypoly, ofile = "", inside = TRUE,
 
 #' @export lasclipCircle
 #' @rdname lasclip
-lasclipCircle = function(x, xcenter, ycenter, radius, ofile = "", inside = TRUE, ...)
+lasclipCircle = function(las, xcenter, ycenter, radius, ...)
 {
-  UseMethod("lasclipCircle", x)
+  UseMethod("lasclipCircle", las)
 }
 
 #' @export
-lasclipCircle.LAS = function(x, xcenter, ycenter, radius, ofile = "", inside = TRUE, ...)
+lasclipCircle.LAS = function(las, xcenter, ycenter, radius, ...)
 {
   assertive::assert_is_numeric(xcenter)
   assertive::assert_is_numeric(ycenter)
   assertive::assert_is_numeric(radius)
   assertive::assert_are_same_length(xcenter, ycenter)
-  if (length(radius) > 1) assertive::assert_are_same_length(xcenter, radius)
-  assertive::assert_is_a_string(ofile)
-  assertive::assert_is_a_bool(inside)
+
+  if (length(radius) > 1)
+    assertive::assert_are_same_length(xcenter, radius)
+  else
+    radius = rep(radius, length(xcenter))
 
   X <- Y <- NULL
 
-  if (length(xcenter) == 1)
+  output = vector(mode = "list", length(xcenter))
+  for (i in 1:length(xcenter))
   {
-    if (inside)
-      return(lasfilter(x, (X-xcenter)^2 + (Y-ycenter)^2 <= radius^2))
-    else
-      return(lasfilter(x, (X-xcenter)^2 + (Y-ycenter)^2 > radius^2))
-  }
-  else
-  {
-    output = vector(mode = "list", length(xcenter))
-    for (i in 1:length(xcenter))
+    roi = lasfilter(las, (X-xcenter[i])^2 + (Y-ycenter[i])^2 <= radius[i]^2)
+    if (is.null(roi))
     {
-      if (inside)
-        output[[i]] = lasfilter(x, (X-xcenter[i])^2 + (Y-ycenter[i])^2 <= radius[i]^2)
-      else
-        output[[i]] = lasfilter(x, (X-xcenter[i])^2 + (Y-ycenter[i])^2 > radius[i]^2)
+      warning(glue::glue("No point found for within disc ({xcenter[i]}, {ycenter[i]}, {radius[i]}). NULL returned."), call. = FALSE)
+      output[i] = list(NULL)
     }
-
-    return(output)
+    else
+      output[[i]] = roi
   }
+
+  if(length(output) == 0)
+    return(NULL)
+  else if (length(output) == 1)
+    return(output[[1]])
+  else
+    return(output)
 }
 
 #' @export
 #' @export
-lasclipCircle.LAScatalog = function(x, xcenter, ycenter, radius, ofile = "", inside = TRUE, ...)
+lasclipCircle.LAScatalog = function(las, xcenter, ycenter, radius, ...)
 {
   assertive::assert_is_numeric(xcenter)
   assertive::assert_is_numeric(ycenter)
   assertive::assert_is_numeric(radius)
   assertive::assert_are_same_length(xcenter, ycenter)
   if (length(radius) > 1) assertive::assert_are_same_length(xcenter, radius)
-  assertive::assert_is_a_string(ofile)
-  assertive::assert_is_a_bool(inside)
 
-  if (!inside) stop("'inside = FALSE' is not available for 'LAScatalog' objects.")
+  xmin   = xcenter - radius
+  xmax   = xcenter + radius
+  ymin   = ycenter - radius
+  ymax   = ycenter + radius
+  bboxes = mapply(raster::extent, xmin, xmax, ymin, ymax)
+  output = catalog_extract(las, bboxes, LIDRCIRCLE, ...)
 
-  if (length(xcenter) == 1)
-    return(catalog_clip_circ(x, xcenter, ycenter, radius, ofile, ...))
+  if (length(output) == 1)
+    return(output[[1]])
   else
-    return(catalog_queries(x, xcenter, ycenter, radius, ...))
+    return(output)
 }
 
+# ========
+# WKT
+# ========
 
-catalog_clip_poly = function(catalog, xpoly, ypoly, ofile, ...)
+lasclipWKT = function(las, wkt, ...)
 {
-  xmin <- min(xpoly)
-  xmax <- max(xpoly)
-  ymin <- min(ypoly)
-  ymax <- max(ypoly)
-  xc   <- (xmax + xmin)/2
-  yc   <- (ymax + ymin)/2
-  w    <- xmax - xmin + 1
-  h    <- ymax - ymin + 1
-
-  cluster  <- catalog_index(catalog, xc, yc, w, h, 0, "ROI")[[1]]
-
-  if (is.null(cluster))
-    return(invisible())
-
-  p = list(...)
-
-  filter = cluster@filter
-  select = "*"
-
-  if(!is.null(p$filter))
-    filter = paste(filter, p$filter)
-
-  if(!is.null(p$select))
-    select = p$select
-
-  header = rlas::read.lasheader(cluster@files[1])
-  data   = rlas:::stream.las(cluster@files, ofile = ofile, select = select, filter = filter)
-
-  header = rlas::read.lasheader(cluster@files[1])
-  data   = rlas:::stream.las_inpoly(cluster@files, xpoly, ypoly, select = select, filter = filter, ofile = ofile)
-
-  if (is.null(data))
-    return (invisible())
-
-  if (nrow(data) == 0)
-    return (invisible())
-
-  return(LAS(data, header))
+  UseMethod("lasclipWKT", las)
 }
 
-catalog_clip_rect = function(catalog, xmin, ymin, xmax, ymax, ofile, ...)
+lasclipWKT.LAS = function(las, wkt, ...)
 {
-  xc <- (xmax + xmin)/2
-  yc <- (ymax + ymin)/2
-  w  <- xmax - xmin
-  h  <- ymax - ymin
+  output = vector(mode = "list", length(wkt))
+  for (i in 1:length(wkt))
+  {
+    roi = lasfilter(las, C_points_in_polygon_wkt(las@data$X, las@data$Y, wkt[i]))
+    if (is.null(roi))
+    {
+      warning(glue::glue("No point found for within {wkt[i]}. NULL returned."), call. = FALSE)
+      output[i] = list(NULL)
+    }
+    else
+      output[[i]] = roi
+  }
 
-  cluster  <- catalog_index(catalog, xc, yc, w, h, 0, "ROI")[[1]]
-
-  if (is.null(cluster))
-    return(invisible())
-
-  p = list(...)
-
-  filter = cluster@filter
-  select = "*"
-
-  if(!is.null(p$filter))
-    filter = paste(filter, p$filter)
-
-  if(!is.null(p$select))
-    select = p$select
-
-  header = rlas::read.lasheader(cluster@files[1])
-  data   = rlas:::stream.las(cluster@files, ofile = ofile, select = select, filter = filter)
-
-  if (is.null(data))
-    return (invisible())
-
-  if (nrow(data) == 0)
-    return (invisible())
-
-  return(LAS(data, header))
+  if(length(output) == 0)
+    return(NULL)
+  else if (length(output) == 1)
+    return(output[[1]])
+  else
+    return(output)
 }
 
-catalog_clip_circ = function(catalog, xcenter, ycenter, radius, ofile, ...)
+lasclipWKT.LAScatalog = function(las, wkt, ...)
 {
-  cluster  <- catalog_index(catalog, xcenter, ycenter, 2*radius, NULL, 0, "ROI")[[1]]
+  bboxes = lapply(wkt, function(string)
+  {
+    spgeom = rgeos::readWKT(string)
+    return(raster::extent(spgeom))
+  })
 
-  if (is.null(cluster))
-    return(invisible())
+  output = catalog_extract(las, bboxes, LIDRRECTANGLE, wkt, ...)
 
-  p = list(...)
-
-  filter = cluster@filter
-  select = "*"
-
-  if(!is.null(p$filter))
-    filter = paste(filter, p$filter)
-
-  if(!is.null(p$select))
-    select = p$select
-
-  header = rlas::read.lasheader(cluster@files[1])
-  data   = rlas:::stream.las(cluster@files, ofile = ofile, select = select, filter = filter)
-
-  if (is.null(data))
-    return (invisible())
-
-  if (nrow(data) == 0)
-    return (invisible())
-
-  return(LAS(data, header))
+  if(length(output) == 0)
+    return(NULL)
+  else if (length(output) == 1)
+    return(output[[1]])
+  else
+    return(output)
 }
 
+# =============
+# GENERIC QUERY
+# =============
+
+catalog_extract = function(ctg, bboxes, shape = LIDRRECTANGLE, wkt = NULL, ...)
+{
+  progress  <- progress(ctg)
+  ncores    <- cores(ctg)
+  stopearly <- stop_early(ctg)
+
+  stopifnot(shape == LIDRRECTANGLE | shape == LIDRCIRCLE)
+
+  if (progress) plot.LAScatalog(ctg, FALSE)
+
+  # Define a function to be passed in cluster_apply
+  extract_query = function(cluster, ...)
+  {
+    if (is.null(cluster)) return(NULL)
+    streamLAS(cluster, ofile = cluster@save, filter_wkt = cluster@wkt, ...)
+  }
+
+  # Find the ROIs in the catalog and return LASclusters
+  clusters <- catalog_index2(ctg, bboxes, shape, 0)
+
+  # Add some useful information in the clusters
+  for (i in 1:length(clusters))
+  {
+    if(!is.null(clusters[[i]]))
+    {
+      if (!is.null(wkt))
+        clusters[[i]]@wkt = wkt[i]
+
+      # if (!is.null(ctg@save))
+      # {
+      #   ID <- i
+      #   XCENTER <- clusters[[i]]@center$x
+      #   XCENTER <- clusters[[i]]@center$y
+      #   XLEFT   <- clusters[[i]]@bbox$xmin
+      #   XRIGHT  <- clusters[[i]]@bbox$xmax
+      #   YBOTTOM <- clusters[[i]]@bbox$ymin
+      #   YTOP    <- clusters[[i]]@bbox$ymax
+      #   format  <- if (ctg@output_options$laz_compression) ".las" else ".las"
+      #   clusters[[i]]@save = paste0(glue::glue(ctg@output_options$output_files), format)
+      # }
+    }
+  }
+
+  output <- cluster_apply(clusters, extract_query, ncores, progress, stopearly, drop_null = FALSE, ...)
+
+  if(length(output) == 0)
+    return(list(NULL))
+
+  # Transfer CRS
+  for (i in 1:length(output))
+  {
+    if (!is.null(output[[i]]))
+    {
+      output[[i]]@crs <- ctg@crs
+
+      # Patch to solves issue #73 waiting for a better solution in issue 2333 in data.table
+      if (ncores > 1)
+        output[[i]]@data <- data.table::alloc.col(output[[i]]@data)
+    }
+  }
+
+  return(output)
+}
+
+# @param ofile character. Path to an output file (only with a \code{LAScatalog} object).
+# If \code{ofile = ""} the result is loaded into R, otherwise the result is written to a
+# file while reading. This is much faster and memory-efficient than loading into R memory first,
+# then writing.