+++ man-roxygen/LAScatalog.R
@@ -1,0 +1,7 @@
+#' @section Working with a \code{LAScatalog}:
+#' In \code{lidR} when the input is a \link[lidR:LAScatalog-class]{LAScatalog} the functions run following
+#' the catalog processing options. The user can modify the processing options using the
+#' \link[lidR:catalog]{available options}.\cr\cr
+#' \code{lidR} supports .lax files that \strong{significantly} improve the computation speed of
+#' spatial queries using a spatial index. Users should really take advange a .lax files but this is
+#' not mandatory.