+++ man-roxygen/param-las.R
@@ -1,0 +1,1 @@
+#' @param las An object of class \code{LAS} or \code{LAScatalog}.